# docs/reference/api-spec.yaml
...
openapi: 3.1.0
info:
  title: SceneFlow Backend
  version: 0.1.0
paths:
...
  /api/admin/users:
    get:
      tags:
      - admin
      summary: List Users
      operationId: list_users_api_admin_users_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Users Api Admin Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - admin
      summary: Create User
      operationId: create_user_api_admin_users_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Payload
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create User Api Admin Users Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/admin/users/{target_user_id}:
    patch:
      tags:
      - admin
      summary: Update User
      operationId: update_user_api_admin_users__target_user_id__patch
      parameters:
      - name: target_user_id
        in: path
        required: true
        schema:
          type: integer
          title: Target User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update User Api Admin Users  Target User Id  Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - admin
      summary: Reset User Password
      operationId: reset_user_password_api_admin_users__target_user_id__post
      parameters:
      - name: target_user_id
        in: path
        required: true
        schema:
          type: integer
          title: Target User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Reset User Password Api Admin Users  Target User Id  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Delete User
      operationId: delete_user_api_admin_users__target_user_id__delete
      parameters:
      - name: target_user_id
        in: path
        required: true
        schema:
          type: integer
          title: Target User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/projects/{project_id}/voices:
    get:
      tags:
      - voices
      summary: List Voices
      operationId: list_voices_api_projects__project_id__voices_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Voices Api Projects  Project Id  Voices Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - voices
      summary: Add Voice
      operationId: add_voice_api_projects__project_id__voices_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVoiceProfileRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Add Voice Api Projects  Project Id  Voices Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}/voices/design:
    post:
      tags:
      - voices
      summary: Design Project Voice
      description: 'Queue a timbre design and bind it to this series when it lands.


        The provider and model come from the project''s audio configuration rather than the

        request. The old form asked the user to type them, which is how a series ended up with

        profiles naming a model no synthesiser here could actually voice.


        The designed voice is also saved to the account''s library, because a timbre that took a

        paid request to produce should be reusable in the next series without paying again.


        Queued rather than awaited: Starlette does not cancel a handler on client disconnect, so

        the stop button could only hang up the browser while this ran on and billed. See

        `app/services/job_worker.py`.'
      operationId: design_project_voice_api_projects__project_id__voices_design_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DesignVoiceProfileRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Design Project Voice Api Projects  Project Id  Voices Design Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/projects/{project_id}/voices/{voice_id}:
    patch:
      tags:
      - voices
      summary: Update Voice
      operationId: update_voice_api_projects__project_id__voices__voice_id__patch
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: voice_id
        in: path
        required: true
        schema:
          type: string
          title: Voice Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVoiceProfileRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Voice Api Projects  Project Id  Voices  Voice Id  Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - voices
      summary: Remove Voice
      operationId: remove_voice_api_projects__project_id__voices__voice_id__delete
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: voice_id
        in: path
        required: true
        schema:
          type: string
          title: Voice Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/voices:
    get:
      tags:
      - user-voices
      summary: List User Voices
      operationId: list_user_voices_api_voices_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List User Voices Api Voices Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/voices/design:
    post:
      tags:
      - user-voices
      summary: Design Voice
      operationId: design_voice_api_voices_design_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Design Voice Api Voices Design Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/voices/{voice_id}/save:
    post:
      tags:
      - user-voices
      summary: Save User Voice
      operationId: save_user_voice_api_voices__voice_id__save_post
      parameters:
      - name: voice_id
        in: path
        required: true
        schema:
          type: string
          title: Voice Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Save User Voice Api Voices  Voice Id  Save Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/voices/{voice_id}:
    delete:
      tags:
      - user-voices
      summary: Delete User Voice
      operationId: delete_user_voice_api_voices__voice_id__delete
      parameters:
      - name: voice_id
        in: path
        required: true
        schema:
          type: string
          title: Voice Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateVoiceProfileRequest:
      properties:
        name:
          type: string
          maxLength: 80
          minLength: 1
          title: Name
        note:
          type: string
          maxLength: 4000
          title: Note
          default: ''
        voiceProvider:
          type: string
          maxLength: 40
          title: Voiceprovider
          default: ''
        voiceModel:
          type: string
          maxLength: 160
          title: Voicemodel
          default: ''
        sampleText:
          type: string
          maxLength: 1000
          title: Sampletext
          default: ''
        orderNum:
          type: integer
          maximum: 9999.0
          minimum: 0.0
          title: Ordernum
          default: 0
      additionalProperties: false
      type: object
      required:
      - name
      title: CreateVoiceProfileRequest
      description: One voice in the show. Provider and model only — credentials stay on the account.
    DesignVoiceProfileRequest:
      properties:
        name:
          type: string
          maxLength: 80
          minLength: 1
          title: Name
        voicePrompt:
          type: string
          maxLength: 1000
          minLength: 1
          title: Voiceprompt
        previewText:
          type: string
          maxLength: 1000
          minLength: 1
          title: Previewtext
        note:
          type: string
          maxLength: 4000
          title: Note
          default: ''
        sampleText:
          type: string
          maxLength: 1000
          title: Sampletext
          default: ''
      additionalProperties: false
      type: object
      required:
      - name
      - voicePrompt
      - previewText
      title: DesignVoiceProfileRequest
      description: 'Design a new timbre and bind it to this series in one step.


        The provider and model come from the project''s audio configuration rather than the

        request: they are an account credential detail, and typing them by hand was how the

        old form let a user create a profile no synthesiser could actually voice.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UpdateVoiceProfileRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 80
            minLength: 1
          - type: 'null'
          title: Name
        note:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Note
        voiceProvider:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Voiceprovider
        voiceModel:
          anyOf:
          - type: string
            maxLength: 160
          - type: 'null'
          title: Voicemodel
        sampleText:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Sampletext
        orderNum:
          anyOf:
          - type: integer
            maximum: 9999.0
            minimum: 0.0
          - type: 'null'
          title: Ordernum
      additionalProperties: false
      type: object
      title: UpdateVoiceProfileRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError